Se separa el código de  actions en fichajes e imputaciones
